Add tree module
